move storage utils into DataStorageRetrieval
